Update eventBranch.m
Looks good for the most part. It's always best to avoid nested for loops but these seem pretty fast, so I wouldn't worry about it too much. A bit of explanation at the top and before any loop is always good if you're writing code that other people will be looking at.