Update eventTree.m
Looking good, changed one for loop to something smaller and more clear, but since your function is taking up most of the time it's probably better to optimize that. Also, if you're ever initializing a variable in a loop over and over to the same thing, it can come before the loop. Headers are always useful